requirements.txt
asgiref==3.6.0
blessed==1.20.0
bpython==0.24
certifi==2022.12.7
charset-normalizer==3.1.0
curtsies==0.4.1
cwcwidth==0.1.8
Django==4.2
django-enumchoicefield==3.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
jedi==0.18.2
packaging==23.0
parso==0.8.3
pluggy==1.0.0
prompt-toolkit==1.0.18
ptpdb==0.17
ptpython==0.33
Pygments==2.15.0
PyJWT==2.6.0
pytest==7.3.0
pytest-django==4.5.2
pytest-testdox==3.0.1
pytz==2023.3
pyxdg==0.28
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
wcwidth==0.2.6
